+++ ChangeLog.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+- added support for getting db log entries via `GET /_admin/log/entries` (ArangoDB v3.8)
 - added support for index estimates (ArangoDB v3.8)
 - added support for ArangoSearch `AQL`, `Pipeline`, `Stopwords`, `GeoJSON`, `GeoPoint` analyzers
 - fixed active failover behavior for the asynchronous driver